v2.3.0: This shiny new preprocessor, with more useful error messages, ooh!

+++ make_exe.py
@@ -1,17 +1,20 @@
 import subprocess
 import os
 import shutil
+import sys
 
 
 try:
     import PyInstaller
 
 except ImportError:
-    subprocess.call(['python', '-m', 'pip', 'install', '-U', 'git+https://github.com/pyinstaller/pyinstaller.git'])
+    import pip
+    pip.main(['install', '-U', 'PyInstaller'])
 
+subprocess.call([sys.executable, '-m', 'PyInstaller', 'zdcode/__main__.py', '-n', 'ZDCode', '-i', 'zdcode.ico', '-F', '--specpath', 'build'])
 
-subprocess.call(['python', '-m', 'PyInstaller', 'zdcode/__main__.py', '-n', 'ZDCode', '-i', 'zdcode.ico', '-F', '--specpath', 'build'])
+if os.path.isfile('./ZDCode.exe'):
+    os.unlink('./ZDCode.exe')
 
-os.unlink('./ZDCode.exe')
 shutil.move('dist/ZDCode.exe', './ZDCode.exe')
 shutil.rmtree('dist')
